qdrant-discovery/requirements.txt
openai==1.74.0
pandas==2.2.3
python-dotenv==1.1.0
qdrant-client==1.13.3
honeyhive==0.2.45
pyarrow==19.0.1
